Trunk: Merged ticket393 at r41745 to the trunk.
Recent tests on Zeus with NAMRR and GSI with the namelist option of .use_reflectivity=.true. has resulted in excessively long running jobs which often exceed their allotted walltime. Unfortunately, simply increasing the walltime for these jobs is not a fix since the runtimes were somewhat erratic.

A system tool was used by Yusong Wang, the high performance computing application specialist on Zeus, to determine the cause of these slow downs. The result of this analysis made it clear that GSI repeatedly attempts to read the microphysics lookup file for the Ferrier scheme (eta_micro_lookup.dat), thus causing a very heavy load on the meta data server on Zeus. Jacob found that read_nems_nmmb_guess (resides in read_wrf_nmm_guess.F90) calls cloud_calc (resides in cloud_efr_mod.f90) to calculate the hydrometeor mixing ratios from the condensate and rain/ice fraction fields from the model. However, inside cloud_calc there is a call to EFFRDS, which uses the inquire statement for every gridpoint and with, what appears to be, every processor (see the call to EFFRDS from cloud_calc).

Based upon additional discussion with developers Yanqiu Zhu and Shun Liu, the call to this inquire statement could be moved. Furthermore, it was noted that for the non-variational cloud analysis routines the lookup table is not used and instead a 95:5 ratio is used for the partitioning of ice hydrometeors.

This update changes how the inquire statement is invoked, with the intent on improving GSI's general performance.

Please note the following updates/bug fixes as well:

 * The anavinfo_gsdcld4nmmb anavinfo fix file was updated to be correctly used with the current GSI trunk.
 * A bug fix was added to update_guess, to allow the NEMS NMMB to update when use_reflectively is set to true.
 * A bug was corrected in wrwrfnmma to correctly write out cloud analysis variables.
 * In the radiance monitoring package, the DO_ARCHIVE default was reset to 0 (off), and the HPSS_DIR and ARCHIVE_DIR location values were set to empty. Normally Radmon users won't have need to do any archiving. The DO_ARCHIVE value was inadvertently set to 1 (on) in the last RadMon release. 
 * Corrected a bug in read_gfs_ozone_for_regional that was causing the NEMS NMM-B to not read in ozone data.


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/util@42234 281fb692-3a18-0410-826d-9e847983f365