Trunk: Merged HyEK_NSST branch at r50979 to the trunk. This update includes changes from four development branches:
Ticket #386 - The NSST (Near-Surface Sea Temperature) has been incorporated and tested with 2011 NCEP GFS version
Ticket #413 - Add AVHRR wind data assimilation to the GSI
Ticket #424 - The assignment of the CRTM wind_direction in the crtm_interface module needs to be modified/corrected.
 * For example for first guess wind components of (u=2m/s,v=2m/s) the wind direction will be the same as for wind component inputs of (u=-2m/s,v=-2m/s). This issue goes back a few years, so it was still relevant for older CRTM releases (i.e application of FASTEM1 etc.)
Ticket #429 - Work is underway to develop a 4D Ensemble-Var capability for implementation in the NCEP GDAS. This ticket is being opened to record the GSI and EnKF changes necessary for the 4D-EnVar implementation.

Finally, an emergency bug fix was added to general_read_gfsatm.f90, increasing the number of fields (nflds) from 5 to 6.

Please keep in mind that these changes do affect results.


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/util@51410 281fb692-3a18-0410-826d-9e847983f365