Trunk: Merged gsi_nsst_update branch at r66047 to the trunk.  This update is an inclusion of two tickets and a regional bug fix:
#453:

A few NSST updates are addressed in this ticket:

 * Move the interpolation of Tf (foundation temperature) analysis increment (from the analysis grids to ensemble grids) to recenter step of EnKF.
 * Modify to enable to provide the SST, which is used at the forecasting mode, in the form of vertical mean or at a specific depth, based the T-Profile produced by NSST Model (NSSTM).
 * Modify to read nst file in the same way as for sfc and sig files, which was update at the end of 2014 by John.
 * Deallocate the nst full resolution arrays earlier to save memory usage.
 * Modify the real variables read from sfc and nst files to be single precision (from double precision) to save memory usage.
 * Modify to make it work when nst_gsi = 0 and nsstbufr data file exists.

#522:

Porting the GSI and EnKF onto the new Cray machine. This ticket will be updated with pertinent information relating to necessary environment variables and when updates have been made to the code.

Bug Fix:

There was a operational NDAS failure on 20151126 12 z. It was caused by a radiosonde report that had more than 255 levels. A bug in subroutine sonde_ext in file read_prepbufr.f90 was found. The check to prevent array out of bound was in place but after printing out an error message the program was stopped.

    if(ll>255)then

        write(6,*)'error in SONDE_EXT levs > 255'
        STOP

    endif

A fix to return to the calling routine which would allow the analysis to continue was included in this branch.

    if(ll>255)then

        write(6,*)'error in SONDE_EXT levs > 255'
        RETURN

    endif


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/util@66575 281fb692-3a18-0410-826d-9e847983f365