Trunk: Merged NEBUDA_CAWV_AMV branch at r73604 to the trunk.  This update adds the ability to assimilate Clear Air Water Vapor AMVs to the GSI.  This effort is to evaluate the use of clear air water vapor (CAWV) AMVs beginning with GOES-13 and GOES-15. Quality control, observational error, spatial and temporal thinning will be investigated. 2-month simulations will be completed to determine impact on the GFS. These AMVs are available from almost all of the geostationary satellites and are already present in the NCEP database. Testing and implementation of this additional AMV type could be accomplished with minimal GSI software modifications. We propose to develop quality control procedures and assimilation techniques for these clear air water vapor AMVs starting with the current 3D+ENKF-VAR system and switch to the developmental 4D+ENKF-VAR system when it is appropriate to use.
The tropics consistently have the greatest wind errors, both in the model and the cloud motion AMVs. This is usually attributed to the localized deep convection influencing both the model winds and the generation of the AMVs. The clear air water vapor AMVs track water vapor image features instead of clouds making them independent of cloud dynamics. One concern with CAWV AMVs is the depth of the layer that is tracked which is dependent on the distribution of the water vapor in the column. The goal is to maximize the positive impact of the data in a region of higher model error and limited operational wind data.

For more details, please see ticket #506.


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/util@74194 281fb692-3a18-0410-826d-9e847983f365